==> src/S01-D-NZV.R <==
# > nearZeroVar(df_train)
# [1] 14 18
# > df_train[,c(14,18)]
# # A tibble: 2,660,252 x 2
# neg_item_cnt_day trc_lgl
# <lgl>            <lgl>
#   1 FALSE            FALSE
# 2 FALSE            FALSE
# 3 FALSE            FALSE
# 4 FALSE            FALSE
# 5 FALSE            FALSE
# 6 FALSE            FALSE
# 7 FALSE            FALSE
# 8 FALSE            FALSE
# 9 FALSE            FALSE
# 10 FALSE            FALSE
# # ... with 2,660,242 more rows
# >
